Renamed to pkg:snippets
