+++ pkgs/snippets/.gitignore
