+++ pkgs/snippets/LICENSE
