+++ pkgs/snippets/analysis_options.yaml
