+++ pkgs/snippets/lib/src/extract.dart
