+++ pkgs/snippets/lib/src/inject.dart
