+++ pkgs/snippets/lib/src/transform.dart
