+++ pkgs/snippets/lib/src/update.dart
