+++ pkgs/snippets/test/extractor_test.dart
@@ -2,8 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:excerpter/src/extract.dart';
 import 'package:path/path.dart' as path;
+import 'package:snippets/src/extract.dart';
 import 'package:test/test.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
