+++ pkgs/snippets/test/injection_test.dart
@@ -2,9 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:excerpter/excerpter.dart';
-import 'package:excerpter/src/inject.dart';
 import 'package:path/path.dart' as path;
+import 'package:snippets/snippets.dart';
+import 'package:snippets/src/inject.dart';
 import 'package:test/test.dart';
 import 'package:test_descriptor/test_descriptor.dart' as d;
