+++ pkgs/snippets/test/transform_test.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:excerpter/excerpter.dart';
+import 'package:snippets/snippets.dart';
 import 'package:test/test.dart';
 
 void main() {
