+++ pkgs/snippets/test_data/example/dartdoc.dart
